refactor: Move `execFile` wrapper to `executeCommand.ts` as `execute`
